creating and moving elastic search helper to its own directory
